+++ experiments/integrated/global/global_soil_canopy.jl
@@ -105,9 +105,10 @@ conductance_args = (; parameters = Canopy.MedlynConductanceParameters(FT; g1))
 photosynthesis_args =
     (; parameters = Canopy.FarquharParameters(FT, is_c3; Vcmax25 = Vcmax25))
 # Set up plant hydraulics
-modis_lai_ncdata_path = ClimaLand.Artifacts.modis_lai_single_year_path(;
+modis_lai_ncdata_path = ClimaLand.Artifacts.modis_lai_multiyear_paths(;
     context = nothing,
-    year = Dates.year(Second(t0) + start_date),
+    start_date = start_date + Second(t0),
+    end_date = start_date + Second(t0) + Second(tf),
 )
 LAIfunction = ClimaLand.prescribed_lai_modis(
     modis_lai_ncdata_path,